update-linux-image-amd64.sh - updated nvidia flags for no interaction

+++ bash/nvidia-drivers/update-linux-image-amd64.sh
@@ -2,17 +2,27 @@
 
 VERSION="580.126.09"
 
-# update linux-image-amd64
-#apt upgrade -y linux-image-amd64
-#apt autoremove -y
-#sleep 3
+if ! /usr/bin/nvidia-smi > /dev/null ; then
+  # update linux-image-amd64
+  # apt upgrade -y linux-image-amd64
+  # apt autoremove -y
+  # sleep 3
 
-# ensure headers are installed
-apt update && apt install -y build-essential linux-headers-$(uname -r)
-sleep 3
+  # ensure headers are installed
+  apt update && apt install -y build-essential linux-headers-$(uname -r)
+  sleep 3
 
-# install driver and reboot
-./NVIDIA-Linux-x86_64-${VERSION}.run --dkms --no-x-check --no-nouveau-check
-sleep 3
-reboot
+  # install driver and reboot
+#  ./NVIDIA-Linux-x86_64-${VERSION}.run --dkms --no-x-check --no-nouveau-check
+  ./NVIDIA-Linux-x86_64-${VERSION}.run \
+	  --no-x-check \
+	  --no-nouveau-check \
+	  --ui=none \
+	  --install-compat32-libs \
+	  --dkms \
+	  --rebuild-initramfs \
+	  --no-nvidia-config
+  sleep 3
+  reboot
+fi
